+++ e2e-tests/tests/viewport-menu.spec.ts
@@ -2,12 +2,14 @@ import { studioTest } from "./infra/studioTest.js";
 
 studioTest("can set and reset Viewport", async ({ studioPage }) => {
   await studioPage.switchPage("LocationPage");
-  await studioPage.openViewportMenu();
-  await studioPage.takePageScreenshotAfterImgRender();
-  await studioPage.setViewport("Galaxy Z Flip5 Folded");
-  await studioPage.openViewportMenu();
-  await studioPage.takePageScreenshotAfterImgRender();
-  await studioPage.setViewport("Reset Viewport");
-  await studioPage.openViewportMenu();
-  await studioPage.takePageScreenshotAfterImgRender();
+  const viewportMenuSection = studioPage.viewportMenuSection;
+
+  await viewportMenuSection.openViewportMenu();
+  await viewportMenuSection.takeScreenshot();
+  await viewportMenuSection.setViewport("Galaxy Z Flip5 Folded");
+  await viewportMenuSection.openViewportMenu();
+  await viewportMenuSection.takeScreenshot();
+  await viewportMenuSection.setViewport("Reset Viewport");
+  await viewportMenuSection.openViewportMenu();
+  await viewportMenuSection.takeScreenshot();
 });
